requirements.txt
telethon
cryptg
pycryptodome
requests
pysocks
hjson
bs4
lxml
yarl